fix: move section archive to hackaton team
